requirements.txt
Django==4.1.7
psycopg2-binary==2.9.5
crispy-bootstrap5==0.7
Pillow==10.0.0
djangorestframework==3.14.0
scrapy==2.8.0
redis==4.5.4

pytest==7.2.2
pytest-django>=4.5.2

django-rq == 2.7.0

flake8 >= 6.0.0
isort>=5.12.0
black>=23.3.0

django-crispy-forms==2.0
django-dotenv==1.4.2
django-allauth
channels==4.0.0
django-ckeditor==6.5.1
django-js-asset==2.0.0
django-cleanup==8.0.0
django-braces==1.15.0
django-taggit==4.0.0


notebook==6.5.4


python-dotenv~=1.0.0